pyproject.toml
[project]
name = "python-bootcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "ipython>=9.10.0",
    "openai>=2.15.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
